+++ windows/flutter/generated_plugins.cmake
@@ -5,7 +5,6 @@
 list(APPEND FLUTTER_PLUGIN_LIST
   app_links
   connectivity_plus
-  desktop_drop
   file_selector_windows
   firebase_core
   flutter_custom_cursor